+++ src/types/iri/absolute.rs
@@ -102,6 +102,11 @@ impl AbsoluteIriStr {
         debug_assert_eq!(validate(s), Ok(()));
         Self::new_always_unchecked(s)
     }
+
+    /// Returns `&str`.
+    pub fn as_str(&self) -> &str {
+        self.as_ref()
+    }
 }
 
 impl std::ops::Deref for AbsoluteIriStr {
@@ -120,7 +125,7 @@ impl fmt::Display for AbsoluteIriString {
 
 impl fmt::Display for &AbsoluteIriStr {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        AsRef::<str>::as_ref(self).fmt(f)
+        f.write_str(self.as_str())
     }
 }